README: Expanded on top level logger configurations
Intended top-level usage (<binary-name> is the library user's
application):

  $ <binary-name> -help
  Usage of <binary-name>:
    -log-dir string
          Write log files in this directory.
    -log-to-stderr
          Log to standard error.
    -log-level (info|debug|warn|error|fatal)
          Log level for logs emitted.
    -log-filter value
          Comma-separated list of pattern:N settings for file-filtered
          logging.

  $ <binary-name> -log-level info \
                  -log-dir /path/to/dir \
                  -log-to-stderr \
                  -log-filter f.go:warn,g/h/*.go:debug

The flags (or some variation thereof) will have to be
provided/registered by the library user, this library simply provides
the hooks the user can then use to configure the logger as needed. These
hooks can be invoked at runtime (in fact, we explicitly avoid init time
hooks and global loggers). What this means is that if needed, a running
service could opt-in to provide open endpoints to accept logger
reconfigurations (via RPCs or otherwise).