Flesh out LogRotationWriter, supporting log rotation
LogRotationWriter returns an io.Writer that internally operates off the
specified directory where writes are written out to rotating files,
thresholded at the specified size in bytes. Within the directory a symlink
is generated pointing to the most recently created log file.

The file names generated for the logs are of the form
<program>.<host>.<user>.<year>-<month>-<day>.<hour>:<minute>:<second>.<millisecond>.<pid>.log
An example: logger.irfansharif-macbook.irfansharif.2018-04-10.22:43:54.717.7989.log